refactor(ios/tts): split PocketTTS model manager responsibilities
- move PocketTTS install and repair sequencing into a dedicated install lifecycle extension
- move PocketTTS staging, manifest, download, and filesystem helpers into a support extension
- update the iOS code map and engineering notes to document the new manager file split
